Moving Documentation and Integration to dedicated entrypoints. (#815)
* Moving Documentation and Integration to dedicated entrypoints.

* Readme: fix imports.
